+++ changelog.md
@@ -11,6 +11,7 @@
 - [command/basic] fix bug: sync taxes before sync products @maramalshen
 - [command/inventory-adjustment] add new Command: adjust inventory from Repzo to QuickBooks @maramalshen
 - [command/inactive_items] add new Command: sync inactive items from QuickBooks to QuickBooks @maramalshen
+- [command/inactive_customers] add new Command: sync inactive Customers from QuickBooks to QuickBooks @maramalshen
 
 ### Changed